Convert users functionality to typescript Add parameter type annotations and return type annotations to functions
