--- src/controllers/User.controller.js
@@ -1,27 +1,0 @@
-import { Router } from 'express';
-
-import { restrictTo } from '../middlewares/Auth.middleware.js';
-
-import { catchAsync } from '../utils/CatchAsync.util.js';
-
-import { UserService } from '../services/User.service.js';
-
-export const UserController = Router();
-
-UserController.get('/', restrictTo('ADMIN'), async (req, res, next) => {
-    const result = await UserService.getAllUsers();
-
-    res.status(result.status).send(result.data);
-});
-
-UserController.get(
-    '/:userId',
-    restrictTo('USER', 'ADMIN'),
-    catchAsync(async (req, res, next) => {
-        const { userId } = req.params;
-
-        const result = await UserService.getUser(userId);
-
-        res.status(result.status).send(result.data);
-    })
-);
